requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.5.0
pandas>=2.1.0

# ── Groq LLM ─────────────────────────────────────────────────────────────────
groq>=0.9.0

# ── HTTP / async ─────────────────────────────────────────────────────────────
aiohttp>=3.9.0
requests>=2.31.0

# ── Web scraping ─────────────────────────────────────────────────────────────
beautifulsoup4>=4.12.0
lxml>=4.9.0
duckduckgo-search>=6.1.0

# ── Dataset sources ───────────────────────────────────────────────────────────
huggingface_hub>=0.23.0
kaggle>=1.6.0

# ── Image downloading ─────────────────────────────────────────────────────────
icrawler>=0.6.6
Pillow>=10.0.0

# ── Logging ───────────────────────────────────────────────────────────────────
rich>=13.7.0
